[TF FE][Tokenizers] Avoid dependency from TF FE in tokenizers
All required routines (Variable, HashTable) are moved to common FE API
