+++ src/frontends/common/src/hash_table.cpp
@@ -1,0 +1,9 @@
+// Copyright (C) 2018-2024 Intel Corporation
+// SPDX-License-Identifier: Apache-2.0
+//
+
+#include "openvino/frontend/hash_table.hpp"
+
+using namespace ov::frontend;
+
+HashTable::~HashTable(){};
