+++ src/frontends/common/src/variable.cpp
@@ -1,0 +1,9 @@
+// Copyright (C) 2018-2024 Intel Corporation
+// SPDX-License-Identifier: Apache-2.0
+//
+
+#include "openvino/frontend/variable.hpp"
+
+using namespace ov::frontend;
+
+Variable::~Variable(){};
